// src/app/models/product.model.ts
export interface Product {
  id:          string;
  title:       string;
  price:       number;
  description: string;
  category:    string;
  image:       string;
}
// export interface Product {
//   id:          number;
//   title:       string;
//   price:       number;
//   description: string;
//   category:    string;
//   image:       string;
//   rating:      Rating;
// }